requirements.txt
streamlit
pandas
python-dotenv
openai
requests
pymupdf
pyarrow
gitpython
tqdm
altair
